Move storage manager to managers folder
