+++ src/balloon.ts
@@ -1,5 +1,5 @@
 import browser from 'webextension-polyfill';
-import storage from '@/storage';
+import storage from '@/managers/storage';
 import { getBalloonContainer, random, sendMessage } from '@/utils';
 
 export const balloonResourceLocation = browser.runtime.getURL(
