+++ src/popup/components/forms/DeleteUser.tsx
@@ -14,7 +14,7 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
-import storage from '@/storage';
+import storage from '@/managers/storage';
 import { User } from '@/const';
 import remote from '@/remote';
