+++ src/popup/components/forms/UpdateUser.tsx
@@ -15,7 +15,7 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
-import storage from '@/storage';
+import storage from '@/managers/storage';
 import { User } from '@/const';
 import remote from '@/remote';
