+++ src/utils.ts
@@ -2,7 +2,7 @@ import browser from 'webextension-polyfill';
 import { Message, BalloonContainerId } from '@/const';
 import { type ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
-import storage from '@/storage';
+import storage from '@/managers/storage';
 import log from 'loglevel';
 
 export function cn(...inputs: ClassValue[]) {
